+++ src/lib/components/MapView/AzimuthsCircle/AzimuthsCircle.svelte
@@ -1,53 +1,68 @@
 <script lang="ts">
 	import type { GetMoonPositionResult, GetSunPositionResult } from "suncalc";
 	import { RADIUS, SVG_CENTER, SVG_SIZE } from "./svgconsts";
-	import { getSectionColor, SourceNameEnum } from "../sectionsConfig";
 	import AzimuthLine from "./AzimuthLine.svelte";
 	import type { AzimuthSection, GeoAzimuths } from "../getGeoAzimuths";
 	import CircleSection from "./CircleSection.svelte";
+	import { SunAzimutIdEnum, getAzimuthSettingsById, DEFAULT_COLOR } from "../sectionsConfig";
 
 	interface Props {
 		sunPos: GetSunPositionResult;
 		moonPos: GetMoonPositionResult;
 		geoPos: GeoAzimuths;
+		highlightedId: SunAzimutIdEnum[];
+		onHighlightedId: (a: SunAzimutIdEnum | undefined) => void;
 	}
 
-	let { sunPos, moonPos, geoPos }: Props = $props();
+	let { sunPos, moonPos, geoPos, highlightedId, onHighlightedId }: Props = $props();
 </script>
 
-{#snippet circleSection(sectionName: SourceNameEnum, coords: AzimuthSection | undefined)}
+{#snippet circleSection(id: SunAzimutIdEnum, coords: AzimuthSection | undefined)}
 	{#if coords}
-		<CircleSection color={getSectionColor(sectionName)} {...coords} />
+		<CircleSection
+			{...getAzimuthSettingsById(id)}
+			{...coords}
+			highlighted={highlightedId.includes(id)}
+			onmouseover={() => onHighlightedId(id)}
+			onmouseleave={() => onHighlightedId(undefined)}
+		/>
 	{/if}
 {/snippet}
 
 <svg xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 {SVG_SIZE} {SVG_SIZE}">
-	<circle
-		cx={SVG_CENTER}
-		cy={SVG_CENTER}
-		r={RADIUS}
-		style:stroke={getSectionColor(SourceNameEnum.Base)}
-	/>
+	<circle cx={SVG_CENTER} cy={SVG_CENTER} r={RADIUS} style:stroke={DEFAULT_COLOR} />
 
-	{@render circleSection(SourceNameEnum.CivilTwilight, geoPos.morningCivilTwilightSection)}
-	{@render circleSection(SourceNameEnum.CivilTwilight, geoPos.eveningCivilTwilightSection)}
+	{@render circleSection(SunAzimutIdEnum.CivilTwilightMorning, geoPos.morningCivilTwilightSection)}
+	{@render circleSection(SunAzimutIdEnum.CivilTwilightEvening, geoPos.eveningCivilTwilightSection)}
 
-	{@render circleSection(SourceNameEnum.GoldenHour, geoPos.morningGoldenHourSection)}
-	{@render circleSection(SourceNameEnum.GoldenHour, geoPos.eveningGoldenHourSection)}
+	{@render circleSection(SunAzimutIdEnum.GoldenHourMorning, geoPos.morningGoldenHourSection)}
+	{@render circleSection(SunAzimutIdEnum.GoldenHourEvening, geoPos.eveningGoldenHourSection)}
 
-	{@render circleSection(SourceNameEnum.Sunrise, geoPos.sunriseSection)}
-	{@render circleSection(SourceNameEnum.Sunset, geoPos.sunsetSection)}
+	{@render circleSection(SunAzimutIdEnum.Sunrise, geoPos.sunriseSection)}
+	{@render circleSection(SunAzimutIdEnum.Sunset, geoPos.sunsetSection)}
 
-	<AzimuthLine stroke={getSectionColor(SourceNameEnum.NoonMarker)} azimuth={geoPos.noon} />
 	<AzimuthLine
-		stroke={getSectionColor(SourceNameEnum.MoonMarker)}
+		{...getAzimuthSettingsById(SunAzimutIdEnum.NoonMarker)}
+		azimuth={geoPos.noon}
+		highlighted={highlightedId.includes(SunAzimutIdEnum.NoonMarker)}
+		onmouseover={() => onHighlightedId(SunAzimutIdEnum.NoonMarker)}
+		onmouseleave={() => onHighlightedId(undefined)}
+	/>
+	<AzimuthLine
+		{...getAzimuthSettingsById(SunAzimutIdEnum.MoonMarker)}
 		azimuth={moonPos.azimuth}
 		dashed={moonPos.altitude < 0}
+		highlighted={highlightedId.includes(SunAzimutIdEnum.MoonMarker)}
+		onmouseover={() => onHighlightedId(SunAzimutIdEnum.MoonMarker)}
+		onmouseleave={() => onHighlightedId(undefined)}
 	/>
 	<AzimuthLine
-		stroke={getSectionColor(SourceNameEnum.SunMarker)}
+		{...getAzimuthSettingsById(SunAzimutIdEnum.SunMarker)}
 		azimuth={sunPos.azimuth}
 		dashed={sunPos.altitude < 0}
+		highlighted={highlightedId.includes(SunAzimutIdEnum.SunMarker)}
+		onmouseover={() => onHighlightedId(SunAzimutIdEnum.SunMarker)}
+		onmouseleave={() => onHighlightedId(undefined)}
 	/>
 </svg>