+++ src/lib/components/MapView/AzimuthsCircle/CircleSection.svelte
@@ -3,14 +3,29 @@
 	import { RADIUS, SVG_CENTER } from "./svgconsts";
 
 	interface Props {
+		/** Section label*/
+		name: string;
 		/** Start azimuth in radians */
 		start: number;
 		/** End azimuth in radians */
 		end: number;
 		/** rgb color hexvalue, without alpha */
 		color: string;
+
+		highlighted?: boolean;
+
+		onmouseover?: () => void | undefined;
+		onmouseleave?: () => void | undefined;
 	}
-	let { start: startAzimuth, end: endAzimuth, color }: Props = $props();
+	let {
+		name,
+		start: startAzimuth,
+		end: endAzimuth,
+		color,
+		highlighted,
+		onmouseleave,
+		onmouseover,
+	}: Props = $props();
 
 	const angleStart = $derived(azimuthRadToUnitCircleRad(startAzimuth));
 	const angleEnd = $derived(azimuthRadToUnitCircleRad(endAzimuth));
@@ -29,6 +44,9 @@
 </script>
 
 <path
+	role="img"
+	aria-label={name}
+	class:highlighted
 	d="M {SVG_CENTER},{SVG_CENTER}
   L {arcStartX},{arcStartY}
   A {RADIUS} {RADIUS} 0 {largeArcFlag} {sweepFlag} {arcEndX} {arcEndY}
@@ -36,10 +54,21 @@
 	stroke={color}
 	fill={color}
 	fill-opacity="0.5"
+	onfocus={onmouseover}
+	{onmouseleave}
+	{onmouseover}
 />
 
 <style>
 	path {
 		stroke-width: 5px;
+	}
+
+	path:focus {
+		outline: 0;
+	}
+
+	.highlighted {
+		filter: drop-shadow(0 0 2px blue);
 	}
 </style>